+++ packages/styleguide/stories/Organisms/examples.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable local-rules/gamut-import-paths */
-import { Box, GridForm } from '@codecademy/gamut/src';
+import { Box, GridForm, Markdown } from '@codecademy/gamut/src';
 import React from 'react';
 
 import { ColorModeExampleWrapper } from '../Foundations/ColorMode/examples';
@@ -64,6 +64,12 @@ export const ColorModeExample = () => {
                   validation: {
                     required: 'You gotta pick one!',
                   },
+                  tooltip: {
+                    children: <Markdown text="There can only be one." />,
+                    id: 'select-tooltip',
+                    alignment: 'bottom-left',
+                    mode: 'light',
+                  },
                 },
                 {
                   label: 'whats up?',
@@ -81,6 +87,12 @@ export const ColorModeExample = () => {
                   type: 'textarea',
                   validation: {
                     required: 'Please write something about penguins!',
+                  },
+                  tooltip: {
+                    children: <Markdown text="It *is* required ✨" />,
+                    id: 'select-tooltip',
+                    alignment: 'top-right',
+                    mode: 'dark',
                   },
                 },
               ],
@@ -102,23 +114,13 @@ export const ColorModeExample = () => {
               ],
               title: 'hi there... again',
             },
-          ]}
-          onSubmit={() => {}}
-          submit={{
-            contents: 'Submit me',
-            size: 12,
-          }}
-          validation="onSubmit"
-        />
-      </Box>
-      <Box my={16}>
-        <GridForm
-          fields={[
             {
               fields: [
                 {
                   label: 'hi?',
-                  name: 'text02-color-mode',
+                  name: 'text01-color-mode-disabled',
+                  disabled: true,
+
                   size: 4,
                   type: 'text',
                   validation: {
@@ -135,7 +137,9 @@ export const ColorModeExample = () => {
                 },
                 {
                   label: 'Simple select (required)',
-                  name: 'simple-select-color-mode',
+                  name: 'simple-select-color-mode-disabled',
+                  disabled: true,
+
                   options: [
                     '',
                     'One fish',
@@ -149,30 +153,153 @@ export const ColorModeExample = () => {
                     required: 'Please select an option',
                   },
                 },
+                {
+                  label: 'Write a paragraph.',
+                  name: 'paragraph01-color-mode-disabled',
+                  disabled: true,
+
+                  size: 8,
+                  type: 'textarea',
+                  validation: {
+                    required: 'Please write something about penguins!',
+                  },
+                },
+                {
+                  label: 'Preferred Modern Artist',
+                  name: 'artist-color-mode-disabled',
+                  disabled: true,
+
+                  options: [
+                    {
+                      label: 'Cardi B',
+                      value: 'cardi',
+                    },
+                    {
+                      label: 'Nicki Minaj',
+                      value: 'nicki',
+                    },
+                  ],
+                  size: 4,
+                  type: 'radio-group',
+                  validation: {
+                    required: 'You gotta pick one!',
+                  },
+                },
+                {
+                  label: 'whats up?',
+                  name: 'text04--left-section-disabled',
+                  size: 4,
+                  type: 'text',
+                  disabled: true,
+                  validation: {
+                    required: true,
+                  },
+                },
+                {
+                  label: 'Write another long paragraph',
+                  name: 'paragraph02-left-section-disabled',
+                  size: 8,
+                  type: 'textarea',
+                  disabled: true,
+                  validation: {
+                    required: 'Please write something about penguins!',
+                  },
+                },
               ],
               layout: 'left',
-              title: 'another form!',
+              title: 'disabled inputs',
             },
             {
-              as: 'h3',
               fields: [
                 {
-                  label: 'hello....',
-                  name: 'text01-center-section',
-                  size: 5,
+                  label: 'hi?',
+                  name: 'text01-color-mode-error',
+                  customError: 'something',
+
+                  size: 4,
                   type: 'text',
                   validation: {
                     required: true,
                   },
                 },
+                {
+                  label: 'Simple select (required)',
+                  name: 'simple-select-color-mode-error',
+                  customError: 'has',
+
+                  options: [
+                    '',
+                    'One fish',
+                    'Two fish',
+                    'Red fish',
+                    'Blue fish',
+                  ],
+                  size: 4,
+                  type: 'select',
+                  validation: {
+                    required: 'Please select an option',
+                  },
+                },
+                {
+                  label: 'Write a paragraph.',
+                  name: 'paragraph01-color-mode-error',
+                  customError: 'gone',
+
+                  size: 8,
+                  type: 'textarea',
+                  validation: {
+                    required: 'Please write something about penguins!',
+                  },
+                },
+                {
+                  label: 'Preferred Modern Artist',
+                  name: 'artist-color-mode-error',
+                  customError: 'very',
+
+                  options: [
+                    {
+                      label: 'Cardi B',
+                      value: 'cardi',
+                    },
+                    {
+                      label: 'Nicki Minaj',
+                      value: 'nicki',
+                    },
+                  ],
+                  size: 4,
+                  type: 'radio-group',
+                  validation: {
+                    required: 'You gotta pick one!',
+                  },
+                },
+                {
+                  label: 'whats up?',
+                  name: 'text04-error',
+                  size: 4,
+                  type: 'text',
+                  customError: 'very',
+
+                  validation: {
+                    required: true,
+                  },
+                },
+                {
+                  label: 'Write another long paragraph',
+                  name: 'paragraph02-error',
+                  size: 8,
+                  type: 'textarea',
+                  customError: 'wrong',
+                  validation: {
+                    required: 'Please write something about penguins!',
+                  },
+                },
               ],
-              title: 'hi there... again',
+              title: 'something has gone wrong',
             },
           ]}
           onSubmit={() => {}}
           submit={{
-            type: 'cta',
-            contents: 'CTA Button',
+            contents: 'Submit me',
             size: 12,
           }}
           validation="onSubmit"